refactor(gui): extract fixture table tooltip and hover helpers
